# Conflicts:
#	src/server/scripts/Draenor/Auchindoun/auchindoun.hpp
#	src/server/scripts/Draenor/Auchindoun/boss_teronogor.cpp
#	src/server/scripts/Draenor/Auchindoun/instance_auchindon.cpp

+++ src/server/game/Battlefield/Zones/BattlefieldWG.h
@@ -1155,7 +1155,7 @@ struct BfWGGameObjectBuilding
     uint32 m_WorldState;
 
     // eWGGameObjectState
-    uint32 m_State;
+    uint64 m_State;
 
     // Name id for warning text
     uint32 m_NameId;
